parallelize preprocessing dump

+++ fast_wikidata_db/preprocess/preprocess_utils/writer_process.py
@@ -1,9 +1,10 @@
+import ujson
 import shutil
 from tqdm import tqdm
-from multiprocessing import Queue
 from pathlib import Path
+from multiprocessing import Queue
 from typing import Dict, Any, List
-import ujson
+
 
 TABLE_NAMES = [
     'labels', 'descriptions', 'aliases', 'entity_values', 'wikipedia_links', 'entity_rels', 'entity_inv_rels'
@@ -11,14 +12,14 @@ TABLE_NAMES = [
 
 
 class Table:
-    def __init__(self, path: Path, batch_nums: int, table_name: str):
+    def __init__(self, path: Path, batch_nums: int, table_name: str, remove_existing: bool = False):
         self.table_name = table_name
         self.table_dir = path / table_name
         self.batch_nums = batch_nums
 
-        if self.table_dir.exists():
+        if self.table_dir.exists() and remove_existing:
             shutil.rmtree(self.table_dir)
-        self.table_dir.mkdir(parents=True, exist_ok=False)
+        self.table_dir.mkdir(parents=True, exist_ok=True)
 
     def write(self, json_value: List[Dict[str, Any]]):
         for json_obj in json_value:
@@ -28,19 +29,22 @@ class Table:
 
 
 class Writer:
-    def __init__(self, path: Path, batch_nums: int, total_num_lines: int):
-        self.output_tables = {table_name: Table(path, batch_nums, table_name) for table_name in TABLE_NAMES}
-        self.progress_bars = {table_name: tqdm(total=total_num_lines, position=i + 1, desc=f"Writing {table_name}") for i, table_name in enumerate(TABLE_NAMES)}
+    def __init__(self, path: Path, batch_nums: int, verbose: bool, remove_existing: bool = False):
+        self.verbose = verbose
+        self.output_tables = {table_name: Table(path, batch_nums, table_name, remove_existing) for table_name in TABLE_NAMES}
+        if self.verbose:
+            self.progress_bar = tqdm(position=10, desc=f"Writing processed dump ")
 
     def write(self, json_object: Dict[str, Any]):
         for key, value in json_object.items():
             if len(value) > 0:
                 self.output_tables[key].write(value)
-                self.progress_bars[key].update(1)
+                if self.verbose and key == "entity_rels":
+                    self.progress_bar.update(10)
 
 
-def write_data(path: Path, batch_nums: int, total_num_lines: int, outout_queue: Queue):
-    writer = Writer(path, batch_nums, total_num_lines)
+def write_data(path: Path, batch_nums: int, verbose: bool, outout_queue: Queue, remove_existing: bool = False):
+    writer = Writer(path, batch_nums, verbose, remove_existing)
     while True:
         json_object = outout_queue.get()
         if json_object is None:
